Moving all classes to their files
